CPS Term Project
The purpose of this project is to produce software to analyze surface water temperaturerecordings for six lakes (Superior, Michigan, Huron, Erie, Ontario, St. Clair) and to produce areport to help us better understand the computed results. The analysis pays special attentionto the relationship between temperature and seasonality for each of the lakes as well as forthe lakes as a group.